internal: move Deque
Move Deque from btf into internal and export methods. No other code changes.
