+++ btf/types.go
@@ -4,11 +4,11 @@ import (
 	"fmt"
 	"io"
 	"math"
-	"math/bits"
 	"reflect"
 	"strings"
 
 	"github.com/cilium/ebpf/asm"
+	"github.com/cilium/ebpf/internal"
 )
 
 const maxTypeDepth = 32
@@ -678,7 +678,7 @@ type copier map[Type]Type
 
 func (c copier) copy(typ *Type, transform Transformer) {
 	var work typeDeque
-	for t := typ; t != nil; t = work.pop() {
+	for t := typ; t != nil; t = work.Pop() {
 		// *t is the identity of the type.
 		if cpy := c[*t]; cpy != nil {
 			*t = cpy
@@ -696,97 +696,11 @@ func (c copier) copy(typ *Type, transform Transformer) {
 		*t = cpy
 
 		// Mark any nested types for copying.
-		walkType(cpy, work.push)
+		walkType(cpy, work.Push)
 	}
 }
 
-type typeDeque = deque[*Type]
-
-// deque implements a double ended queue.
-type deque[T any] struct {
-	elems       []T
-	read, write uint64
-	mask        uint64
-}
-
-func (dq *deque[T]) empty() bool {
-	return dq.read == dq.write
-}
-
-func (dq *deque[T]) remainingCap() int {
-	return len(dq.elems) - int(dq.write-dq.read)
-}
-
-// push adds an element to the end.
-func (dq *deque[T]) push(e T) {
-	if dq.remainingCap() >= 1 {
-		dq.elems[dq.write&dq.mask] = e
-		dq.write++
-		return
-	}
-
-	elems := dq.linearise(1)
-	elems = append(elems, e)
-
-	dq.elems = elems[:cap(elems)]
-	dq.mask = uint64(cap(elems)) - 1
-	dq.read, dq.write = 0, uint64(len(elems))
-}
-
-// shift returns the first element or the zero value.
-func (dq *deque[T]) shift() T {
-	var zero T
-
-	if dq.empty() {
-		return zero
-	}
-
-	index := dq.read & dq.mask
-	t := dq.elems[index]
-	dq.elems[index] = zero
-	dq.read++
-	return t
-}
-
-// pop returns the last element or the zero value.
-func (dq *deque[T]) pop() T {
-	var zero T
-
-	if dq.empty() {
-		return zero
-	}
-
-	dq.write--
-	index := dq.write & dq.mask
-	t := dq.elems[index]
-	dq.elems[index] = zero
-	return t
-}
-
-// linearise the contents of the deque.
-//
-// The returned slice has space for at least n more elements and has power
-// of two capacity.
-func (dq *deque[T]) linearise(n int) []T {
-	length := dq.write - dq.read
-	need := length + uint64(n)
-	if need < length {
-		panic("overflow")
-	}
-
-	// Round up to the new power of two which is at least 8.
-	// See https://jameshfisher.com/2018/03/30/round-up-power-2/
-	capacity := 1 << (64 - bits.LeadingZeros64(need-1))
-	if capacity < 8 {
-		capacity = 8
-	}
-
-	types := make([]T, length, capacity)
-	pivot := dq.read & dq.mask
-	copied := copy(types, dq.elems[pivot:])
-	copy(types[copied:], dq.elems[:pivot])
-	return types
-}
+type typeDeque = internal.Deque[*Type]
 
 // inflateRawTypes takes a list of raw btf types linked via type IDs, and turns
 // it into a graph of Types connected via pointers.
