+++ btf/types_test.go
@@ -193,83 +193,6 @@ func countChildren(t *testing.T, typ reflect.Type) int {
 	return n
 }
 
-func TestDeque(t *testing.T) {
-	t.Run("pop", func(t *testing.T) {
-		var dq deque[int]
-		dq.push(1)
-		dq.push(2)
-
-		if dq.pop() != 2 {
-			t.Error("Didn't pop 2 first")
-		}
-
-		if dq.pop() != 1 {
-			t.Error("Didn't pop 1 second")
-		}
-
-		if dq.pop() != 0 {
-			t.Error("Didn't pop zero")
-		}
-	})
-
-	t.Run("shift", func(t *testing.T) {
-		var td deque[int]
-		td.push(1)
-		td.push(2)
-
-		if td.shift() != 1 {
-			t.Error("Didn't shift 1 first")
-		}
-
-		if td.shift() != 2 {
-			t.Error("Didn't shift b second")
-		}
-
-		if td.shift() != 0 {
-			t.Error("Didn't shift zero")
-		}
-	})
-
-	t.Run("push", func(t *testing.T) {
-		var td deque[int]
-		td.push(1)
-		td.push(2)
-		td.shift()
-
-		for i := 1; i <= 12; i++ {
-			td.push(i)
-		}
-
-		if td.shift() != 2 {
-			t.Error("Didn't shift 2 first")
-		}
-		for i := 1; i <= 12; i++ {
-			if v := td.shift(); v != i {
-				t.Fatalf("Shifted %d at pos %d", v, i)
-			}
-		}
-	})
-
-	t.Run("linearise", func(t *testing.T) {
-		var td deque[int]
-		td.push(1)
-		td.push(2)
-
-		all := td.linearise(0)
-		if len(all) != 2 {
-			t.Fatal("Expected 2 elements, got", len(all))
-		}
-
-		if cap(all)&(cap(all)-1) != 0 {
-			t.Fatalf("Capacity %d is not a power of two", cap(all))
-		}
-
-		if all[0] != 1 || all[1] != 2 {
-			t.Fatal("Elements don't match")
-		}
-	})
-}
-
 type testFormattableType struct {
 	name  string
 	extra []interface{}
@@ -471,7 +394,7 @@ func BenchmarkWalk(b *testing.B) {
 
 			for i := 0; i < b.N; i++ {
 				var dq typeDeque
-				walkType(typ, dq.push)
+				walkType(typ, dq.Push)
 			}
 		})
 	}
